Extract Teensy into custom UserControl
