+++ AOGConfigOMatic/AOGConfigOMatic.csproj
@@ -60,6 +60,12 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Teensy\TeensyControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Teensy\TeensyControl.Designer.cs">
+      <DependentUpon>TeensyControl.cs</DependentUpon>
+    </Compile>
     <EmbeddedResource Include="frmMain.resx">
       <DependentUpon>frmMain.cs</DependentUpon>
     </EmbeddedResource>
@@ -73,6 +79,9 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
+    <EmbeddedResource Include="Teensy\TeensyControl.resx">
+      <DependentUpon>TeensyControl.cs</DependentUpon>
+    </EmbeddedResource>
     <None Include="Firmwares.csv" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
